Create GUI Module SOL-24 #done
