+++ src/main.cpp
@@ -1,6 +1,7 @@
 #include "components.h"
 #include "graphics.h"
 #include "gui.h"
+#include "modules/gui.h"
 #include "modules/main_menu.h"
 #include "modules/rocket_launch.h"
 #include "spdlog/sinks/basic_file_sink.h"
@@ -64,8 +65,6 @@ void registerSystems(flecs::world &world) {
       .kind(preFramePhase)
       .each(systemEventHandling);
 
-  world.system("New GUI Frame").kind(preFramePhase).run(systemGuiNewFrame);
-
   // Simulator Systems
   world.system<GameResource>("Update Simulation Date")
       .tick_source(game->sim_speed)
@@ -85,9 +84,6 @@ void registerSystems(flecs::world &world) {
       .kind(guiPhase)
       .each(systemDrawSiteWindow);
 
-  // Render Phase
-  world.system("End GUI Frame").kind(preRenderPhase).run(systemGuiEndFrame);
-
   world.system<const Renderer>("Render Begin")
       .term_at(0)
       .singleton()
@@ -100,12 +96,6 @@ void registerSystems(flecs::world &world) {
       .kind(renderPhase)
       .each(systemRenderSprite);
 
-  world.system<const Renderer>("Render GUI")
-      .term_at(0)
-      .singleton()
-      .kind(renderPhase)
-      .each(systemRenderGUI);
-
   world.system<const Renderer>("Render End")
       .term_at(0)
       .singleton()
@@ -115,7 +105,6 @@ void registerSystems(flecs::world &world) {
 
 void prepareGraphics(flecs::world &world) {
   initialiseGraphics(world);
-  initialiseGUI(world);
 
   Texture t = loadTexture("../../textures/solcorp_buildings.png", world);
   t.cols = 1;
@@ -155,6 +144,7 @@ int main(void) {
 
   world.import <MainMenuModule>();
   world.import <RocketLaunchModule>();
+  world.import <GuiModule>();
 
   auto site =
       world.entity("cape_canaveral").set<Site>({"Cape Canaveral", 10, 10});
