+++ src/modules/gui.h
@@ -1,0 +1,8 @@
+#pragma once
+
+#include <flecs.h>
+
+struct GuiModule {
+public:
+  GuiModule(flecs::world &);
+};
